fix:Initialize uv project for repeatable dev
Before this commit, the project's `main.py` could only be run with a
python which already had the necessary dependencies installed.

This change adds `pyproject.toml` and `.python-version` files as
initialized by `uv`. The lock file is committed as well to ensure
different developers resolve the same dependency versions.

The initial dependencies are:
* python 3.13 (with tkinter)
* FreeSimpleGUI 5.2.0.post1
* pymupdf 1.26.3

Further, the README has been updated with steps for using uv to run and
the project.

Finally, a `.gitignore` was added which prevents `__pycache__` folders
from being commited to source.

This improves development across multiple developers by ensuring they
target the same dependencies, and can easily check out and use the
project.

References:
* https://docs.astral.sh/uv/
* https://www.conventionalcommits.org/en/v1.0.0/